requirements.txt
kurigram
tgcrypto
python-dotenv
aiohttp
rich
uvloop; platform_system != 'Windows'
psutil
